refactor: 상품 구매 메뉴 생성 및 클릭 이벤트 view로 이동
